RF: Remove duplication, move code to generate test data into imwarp module.
