+++ dipy/align/imwarp.py
@@ -11,6 +11,7 @@ import dipy.align.vector_fields as vfu
 from dipy.align import floating
 from dipy.align import VerbosityLevels
 from dipy.align import Bunch
+from dipy.data import get_data
 
 RegistrationStages = Bunch(INIT_START=0,
                            INIT_END=1,
@@ -1719,3 +1720,51 @@ class SymmetricDiffeomorphicRegistration(DiffeomorphicRegistration):
         self._optimize()
         self._end_optimizer()
         return self.static_to_ref
+
+
+def get_synthetic_warped_circle(nslices):
+    """
+    Create synthetic data to test registration algorithms
+
+    Parameters
+    ----------
+    nslices : int
+        The number of slices in the volume
+
+    Returns
+    -------
+    circle_3d : 3D array, shape (64, 64, nslices)
+        A volume with a cylinder
+
+    wcircle_3d : 3D array, shape (64, 64, nslices)
+        A volume with the same cylinder warped through a DiffeomorphicMap
+    """
+    #get a subsampled circle
+    fname_cicle = get_data('reg_o')
+    circle = np.load(fname_cicle)[::4,::4].astype(floating)
+
+    #create a synthetic invertible map and warp the circle
+    d, dinv = vfu.create_harmonic_fields_2d(64, 64, 0.1, 4)
+    d = np.asarray(d, dtype=floating)
+    dinv = np.asarray(dinv, dtype=floating)
+    mapping = DiffeomorphicMap(2, (64, 64))
+    mapping.forward, mapping.backward = d, dinv
+    wcircle = mapping.transform(circle)
+
+    if(nslices == 1):
+        return circle, wcircle
+
+    #normalize and form the 3d by piling slices
+    circle = (circle-circle.min())/(circle.max() - circle.min())
+    circle_3d = np.ndarray(circle.shape + (nslices,), dtype=floating)
+    circle_3d[...] = circle[...,None]
+    circle_3d[...,0] = 0
+    circle_3d[...,-1] = 0
+
+    #do the same with the warped circle
+    wcircle = (wcircle-wcircle.min())/(wcircle.max() - wcircle.min())
+    wcircle_3d = np.ndarray(wcircle.shape + (nslices,), dtype=floating)
+    wcircle_3d[...] = wcircle[...,None]
+    wcircle_3d[...,0] = 0
+    wcircle_3d[...,-1] = 0
+    return circle_3d, wcircle_3d
